pyproject.toml
[project]
name = "langchain-agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.2",
    "langchain-huggingface>=1.2.2",
    "langchain-openai>=1.2.1",
    "langchain-text-splitters>=1.1.2",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "tiktoken>=0.12.0",
]
